backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
sqlalchemy
aiosqlite
pytest
httpx
pytest-asyncio
greenlet
python-jose[cryptography]
passlib
pydantic[email]
python-multipart